Time: 1 ms (64.91%), Space: 40.3 MB (74.53%) - LeetHub

+++ 1929-maximum-value-at-a-given-index-in-a-bounded-array/1929-maximum-value-at-a-given-index-in-a-bounded-array.java
@@ -2,20 +2,20 @@ class Solution {
     public int maxValue(int n, int index, int maxSum) {
         // 不超过 maxSum，数字越小越好
 
-        int leftNum = index;
-        int rightNum = n - leftNum - 1;
+        long leftNum = index;
+        long rightNum = n - leftNum - 1;
 
-        int start = 1;
-        int end = maxSum + 1;
-        int res = -1;
+        long start = 1;
+        long end = maxSum + 1;
+        long res = -1;
         while (start < end) {
-            int candidate = start + (end - start) / 2;
+            long candidate = start + (end - start) / 2;
 
-            int leftPositiveNum = Math.min(leftNum, candidate - 1);
+            long leftPositiveNum = Math.min(leftNum, candidate - 1);
             long leftSum = leftPositiveNum * candidate - (((leftPositiveNum * (leftPositiveNum + 1))) / 2);
             leftSum += leftNum - leftPositiveNum;
             
-            int rightPositiveNum = Math.min(rightNum, candidate - 1);
+            long rightPositiveNum = Math.min(rightNum, candidate - 1);
             long rightSum =  rightPositiveNum * candidate - (((rightPositiveNum * (rightPositiveNum + 1))) / 2);
             rightSum += rightNum - rightPositiveNum;
 
@@ -29,6 +29,6 @@ class Solution {
             }
         }
       
-        return res;
+        return (int)res;
     }
 }
